chore: revert the change to device name in inferece/annotator
