+++ application/ui/src/components/util.ts
@@ -22,22 +22,3 @@ export const getFormatOptions = (taskType: TaskType) => {
 
     return options[taskType];
 };
-
-const formatDeviceMemory = (bytes: number): string => {
-    return `${Math.ceil(bytes / 1024 ** 3)} GB`;
-};
-
-export const createDeviceName = (device: { name: string; index?: number | null; memory?: number | null }): string => {
-    let name = device.name;
-
-    if (device.memory != null) {
-        const memory = formatDeviceMemory(device.memory);
-        name += ` (${memory})`;
-    }
-
-    if (device.index != null) {
-        name += ` [${device.index}]`;
-    }
-
-    return name;
-};
