+++ application/ui/src/features/models/train-model/select-training-device/utils.test.ts
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import type { TrainingDevice } from '../../../../constants/shared-types';
-import { getDefaultTrainingDevice } from './utils';
+import { createDeviceName, getDefaultTrainingDevice } from './utils';
 
 const makeDevice = (overrides: Partial<TrainingDevice> = {}): TrainingDevice => ({
     type: 'cpu',
@@ -72,5 +72,36 @@ describe('getDefaultTrainingDevice', () => {
         const iGPU = makeDevice({ type: 'xpu', name: 'Intel iGPU', memory: 8_589_934_592, index: 1 });
 
         expect(getDefaultTrainingDevice([cpu, arcA770, iGPU])).toBe(arcA770);
+    });
+});
+
+describe('createDeviceName', () => {
+    it('returns just the name when memory and index are both null', () => {
+        expect(createDeviceName(makeDevice({ name: 'CPU', memory: null, index: null }))).toBe('CPU');
+    });
+
+    it('returns just the name when memory and index are undefined', () => {
+        expect(createDeviceName(makeDevice({ name: 'CPU' }))).toBe('CPU');
+    });
+
+    it('appends memory in GB (rounded up) when memory is set', () => {
+        expect(createDeviceName(makeDevice({ name: 'GPU', memory: 8_000_000_000 }))).toBe('GPU (8 GB)');
+    });
+
+    it('rounds up fractional GB values with Math.ceil', () => {
+        // 1 byte over 7 GB → rounds up to 8 GB
+        expect(createDeviceName(makeDevice({ name: 'GPU', memory: 7 * 1024 ** 3 + 1 }))).toBe('GPU (8 GB)');
+    });
+
+    it('appends index in brackets when index is set', () => {
+        expect(createDeviceName(makeDevice({ name: 'XPU', index: 0 }))).toBe('XPU [0]');
+    });
+
+    it('appends memory before index when both are set', () => {
+        expect(createDeviceName(makeDevice({ name: 'XPU', memory: 21_359_386_624, index: 1 }))).toBe('XPU (20 GB) [1]');
+    });
+
+    it('appends memory when memory is 0 (0 != null is true)', () => {
+        expect(createDeviceName(makeDevice({ name: 'CPU', memory: 0 }))).toBe('CPU (0 GB)');
     });
 });
